Version 0.5.8 with 'zucchini approve' subcommand
