+++ bin/zucchini
@@ -8,6 +8,7 @@ require 'lib/config'
 require 'lib/screenshot'
 require 'lib/report'
 require 'lib/feature'
+require 'lib/detector'
 require 'lib/runner'
 require 'lib/generator'
 require 'lib/approver'
