+++ lib/version.rb
@@ -1,3 +1,3 @@
 module Zucchini
-  VERSION = "0.5.7"
+  VERSION = "0.5.8"
 end
