+++ spec/lib/detector_spec.rb
@@ -1,29 +1,29 @@
 require 'spec_helper'
 
-describe Zucchini::Runner do
-  before (:each) do
-    runner.path = "spec/sample_setup/feature_one"
+describe Zucchini::Detector do
+  before do
+    detector.path = "spec/sample_setup/feature_one"
     ENV['ZUCCHINI_DEVICE'] = 'My iDevice'
   end 
    
-  let (:runner) { Zucchini::Runner.new(nil) }
-  
+  let (:detector) { Zucchini::Detector.new(nil) }
+
   describe "execute" do
-    subject { lambda { runner.execute } }
+    subject { lambda { detector.execute } }
     
     context "feature directory doesn't exist" do
-      before { runner.path = "spec/sample_setup/erroneous_feature" }
+      before { detector.path = "spec/sample_setup/erroneous_feature" }
       it     { should raise_error "Directory spec/sample_setup/erroneous_feature does not exist" }
     end
-    
+
     context "device hasn't been found" do
       before { ENV['ZUCCHINI_DEVICE'] = 'My Android Phone' }
       it     { should raise_error "Device not listed in config.yml" }
     end
   end
-  
+
   describe "detect features" do
-    subject { lambda { runner.detect_features(@path) } }
+    subject { lambda { detector.detect_features(@path) } }
     
     context "path to a single feature" do
       before { @path = "spec/sample_setup/feature_one" }
