+++ spec/spec_helper.rb
@@ -6,6 +6,7 @@ require 'lib/config'
 require 'lib/screenshot'
 require 'lib/report'
 require 'lib/feature'
+require 'lib/detector'
 require 'lib/runner'
 require 'lib/generator'
 require 'lib/approver'
